refactor(typescript): migrate JavaScript files to TypeScript with improved type safety
feat(constants): add new constants module for earth and projection calculations
refactor(layer): convert ThreeLayer to TypeScript with enhanced documentation
refactor(transform): update SceneTransform with type annotations and better docs
refactor(creator): rewrite Creator in TypeScript with improved type safety
refactor(camera): convert CameraSync to TypeScript with maplibre-gl types
refactor(utils): migrate Util to TypeScript with proper type definitions
refactor(sun): rewrite Sun in TypeScript with detailed interfaces
refactor(scene): convert MapScene to TypeScript with comprehensive interfaces
build(package): add @types/three and maplibre-gl as peer dependencies
build(eslint): enable TypeScript support in eslint config
style(tsdown): fix formatting in config file
build(tsconfig): add dom to lib for browser API support
