--- .babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "modules": false,
-        "targets": {
-          "browsers": ["> 1%", "last 2 versions", "ie >= 10"]
-        }
-      }
-    ]
-  ],
-  "plugins": ["@babel/plugin-transform-runtime","@babel/plugin-proposal-class-properties"]
-}
