+++ .gitignore
@@ -15,3 +15,4 @@ package-lock.json
 .DS_Store
 .history
 */config.js
+dist
