+++ eslint.config.js
@@ -5,4 +5,5 @@ export default antfu({
     quotes: 'single',
     semi: false,
   },
+  typescript: true,
 })
