--- src/index.js
@@ -1,3 +1,0 @@
-import { MapScene, SceneTransform, Creator, Sun } from './modules'
-
-export { MapScene, SceneTransform, Creator, Sun }
