+++ src/modules/constants/index.ts
@@ -6,11 +6,11 @@ const RAD2DEG = 180 / Math.PI
 const PROJECTION_WORLD_SIZE = WORLD_SIZE / EARTH_CIRCUMFERENCE
 const TILE_SIZE = 512
 export {
+  DEG2RAD,
+  EARTH_CIRCUMFERENCE,
+  EARTH_RADIUS,
+  PROJECTION_WORLD_SIZE,
+  RAD2DEG,
   TILE_SIZE,
   WORLD_SIZE,
-  EARTH_RADIUS,
-  EARTH_CIRCUMFERENCE,
-  DEG2RAD,
-  RAD2DEG,
-  PROJECTION_WORLD_SIZE,
 }
