--- src/modules/index.js
@@ -1,4 +1,0 @@
-export { default as MapScene } from './scene/MapScene'
-export { default as SceneTransform } from './transform/SceneTransform'
-export { default as Sun } from './sun/Sun'
-export { default as Creator } from './creator/Creator'
